pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "fabianfernandez", email = "fabian.fernandeza@sansano.usm.cl"},
]
dependencies = [
    "django>=4.2.3",
    "django-environ>=0.10.0",
    "psycopg>=3.1.9",
    "psycopg2>=2.9.6",
    "djangorestframework>=3.14.0",
    "markdown>=3.5",
    "django-filter>=23.3",
    "django-cors-headers>=4.3.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}
